requirements.txt
Mezzanine==4.2.2
django-crispy-forms==1.6.1
